src/processor/main.py
#!/usr/bin/env python3
"""
Minimal Cloud Run Job for ETL CSV Processing
"""

import os
import json
import logging
from google.cloud import storage

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class ETLProcessor:
    """Minimal ETL processor for Cloud Run Job"""
    
    def __init__(self):
        self.project_id = os.getenv('GCP_PROJECT', 'max-sm-ingest-dev')
        self.region = os.getenv('GCP_REGION', 'us-central1')
        self.bucket_name = os.getenv('STORAGE_BUCKET', 'sm-ingest-dev-8')
        self.storage_client = storage.Client(project=self.project_id)
        
    def process_manifest(self, manifest_data):
        """Process the manifest file and validate CSV files"""
        try:
            logger.info(f"Processing manifest for load_id: {manifest_data.get('load_id')}")
            
            if 'files' not in manifest_data:
                logger.error("Invalid manifest: missing 'files' key")
                return False
                
            files = manifest_data['files']
            logger.info(f"Found {len(files)} files in manifest")
            
            for file_info in files:
                file_name = file_info.get('name', '')
                row_count = file_info.get('rows', 0)
                
                logger.info(f"Processing file: {file_name} (rows: {row_count})")
                    
            logger.info("Manifest processing completed successfully")
            return True
            
        except Exception as e:
            logger.error(f"Error processing manifest: {str(e)}")
            return False
    
    def run(self, load_id=None, partner_id=None, shop_id=None):
        """Main entry point for the Cloud Run Job"""
        try:
            logger.info("Starting ETL Processor Job")
            
            load_id = load_id or os.getenv('LOAD_ID', 'test-load-001')
            partner_id = partner_id or os.getenv('PARTNER_ID', 'test-partner')
            shop_id = shop_id or os.getenv('SHOP_ID', 'test-shop')
                
            logger.info(f"Processing load_id: {load_id}, partner: {partner_id}, shop: {shop_id}")
            
            # Create a sample manifest for testing
            sample_manifest = {
                "load_id": load_id,
                "files": [
                    {"name": "customers.csv", "rows": 100, "sha256": "sample_hash_1"},
                    {"name": "vehicles.csv", "rows": 50, "sha256": "sample_hash_2"},
                    {"name": "invoices.csv", "rows": 0, "sha256": "sample_hash_3"},
                    {"name": "line_items.csv", "rows": 200, "sha256": "sample_hash_4"},
                    {"name": "payments.csv", "rows": 0, "sha256": "sample_hash_5"},
                    {"name": "inventory_parts.csv", "rows": 25, "sha256": "sample_hash_6"},
                    {"name": "suppliers.csv", "rows": 10, "sha256": "sample_hash_7"}
                ]
            }
            
            success = self.process_manifest(sample_manifest)
            return 0 if success else 1
                
        except Exception as e:
            logger.error(f"Fatal error in ETL processor: {str(e)}")
            return 1


def main():
    """Main entry point for the Cloud Run Job"""
    logger.info("ETL Processor Job starting...")
    
    processor = ETLProcessor()
    exit_code = processor.run()
    
    logger.info(f"ETL Processor Job completed with exit code: {exit_code}")
    exit(exit_code)


if __name__ == "__main__":
    main()