src/processor/requirements.txt
google-cloud-storage==2.15.0
google-cloud-run==0.10.0
PyYAML==6.0.2
